pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rocketstocks"
version = "0.5.0"
requires-python = ">=3.12"
dependencies = [
    "APScheduler",
    "discord.py",
    "pandas-datareader",
    "yfinance",
    "pandas-ta",
    "mplfinance",
    "Backtesting",
    "tradingview_screener",
    "newsapi-python",
    "table2ascii",
    "psycopg2-binary",
    "ratelimit",
    "pandas_market_calendars",
    "schwab-py",
    "beautifulsoup4",
    "numpy==1.26.4",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]
